requirements.txt
aiodns==3.0.0
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.2.0
brotlipy==0.7.0
cchardet==2.1.7
cffi==1.14.6
chardet==4.0.0
idna==3.2
maxminddb==2.2.0
maxminddb-geolite2==2018.703
multidict==5.1.0
pycares==4.0.0
pycparser==2.20
python-geoip==1.2
typing-extensions==3.10.0.2
yarl==1.6.3
